Moved BackendConfig into its own file
